Added the THEMIS downloaders in __init__.py
in setup.cfg I added the cdflib dependency.

Wrote the loader in plot_frame.py. Given the
mission, station, and date, the load() function will
attempt to find a matching file locally. If one is not
found, it will attempt to download it. This function
returns a cdflib.CDF object (lazy opening so it is just
a handle).