Wrote the .sav callibration file reader to load in the REGO or THEMIS callibration data. If a file does not exist, it will attempt to download it.
The calibration data is saved to a dictionary and the
longitudes are mapped to -180 to 180 degrees.

test_load.py: I added a few tests to check that the
THEMIS and REGO claibration and image files are
loaded. For test_get_frame(), I test that the retrieved
image time stamp is within 3 seconds of the requested
time stamp.