load.py: Added a check to load_frames that the time_range is ordered.
Also if the YYYY-MM-DD and HOUR is not the same between
time_range[0] and time_range[1], this code will load all of
the hour chunks in the time range. This way if the time range
spans two hours, data will be loaded. Also if you want to make a time
lapse of the aurora over multiple hours, this function supports loading
data from an arbitary number of hours (RAM limited of course).

plot_movie.py: I added a np.all() check that the image pixels are not all
0s (I assume that it is an error in the ASI). I also added a try-catch block
if no data was found in the date_range even though the hour files
exist.