I figured out the orientation problem and updated plot_frame(). I added azel_countors kwarg to plot the azimuth and elevation contours.
I need to plot ax.imshow with origin='lower' instead
of the default 'upper' (upper/lower left corner for the
image origin).

Also both axes in the azimuth and elevation calibration arrays needed
to be flipped, i.e. cal_dict['FULL_AZIMUTH'][::-1, ::-1] and the image
lines up well with the calibration data.

Next: rotate the images in the plot_movie() function, and in
the project_lla_to_skyfield().