Working on the lat/lon masking.
Added many helpfull comments to
pcolormesh_nan(). Now I understand this
function that remaps the lat/lon maps. There are two main steps:

1) All nan values to the left of the first valid value are
reassgned to the first valid value. Likewise, all nan values to the
right of the last valid value are reassigned to it.

2) All nan-filled rows above the first valid row are assigned to the
maximum value in the first row, likewiese for the bottom rows.

Essentially this is a compression of all nan values in the perhiphery
to the center valid values.