Fixed the double slash in the LAMP EMCCD downloader call.
I also found out that requests.get() can potentially download a large
file in order to check server status. I added a timeout in case
the server takes time to respond, so it doesn't hang perpetually.
Also, I added a check if self.url contains the filename to strip it off,
so that the server status executes quickly.